+++ hamonize-agent/debian/changelog
@@ -1,0 +1,5 @@
+hamonize-agent (1.0.0) stable; urgency=low
+
+  * Initial release. 1.0.0
+
+ -- HamoniKR <pkg@hamonikr.org>  Wed, 26 May 2021 15:51:27 +0900
