+++ hamonize-agent/debian/compat
@@ -1,0 +1,1 @@
+10
