+++ hamonize-agent/debian/control
@@ -1,0 +1,13 @@
+Source: hamonize-agent
+Section: unknown
+Priority: optional
+Maintainer: HamoniKR Team <hamonikr-packages@googlegroups.com>
+Build-Depends: debhelper (>=11~)
+Standards-Version: 4.1.4
+Homepage: https://hamonikr.org
+
+Package: hamonize-agent
+Architecture: any
+Depends: ${misc:Depends}, ${shlibs:Depends}
+Description: Agent program that manages pc in Harmonize projects.
+ 하모나이즈 프로젝트에서 pc를 관리하는 에이전트 프로그램입니다.
