+++ hamonize-agent/debian/install
@@ -1,0 +1,3 @@
+usr
+etc
+var
