+++ hamonize-agent/debian/postinst
@@ -1,0 +1,12 @@
+#!/bin/bash
+# make uuid 
+uuid=`cat /etc/machine-id`
+echo "$uuid" >> /etc/hamonize/uuid
+
+# install jq
+sudo apt install jq
+
+# hamonize-agent service start 
+sudo systemctl daemon-reload
+sudo systemctl enable hamonize-agent
+sudo systemctl start hamonize-agent
