+++ hamonize-agent/debian/source/format
@@ -1,0 +1,1 @@
+3.0 (native)
