+++ hamonize-agent/etc/hamonize/progrm/progrm.hm
@@ -1,0 +1,1 @@
+{"DEL":"filezilla"}
