+++ hamonize-agent/etc/hamonize/propertiesJob/propertiesInfo.hm
@@ -1,0 +1,2 @@
+CENTERURL=192.168.0.210:8082
+APTURL=update.skir.kr
