+++ hamonize-agent/etc/hamonize/siteblock/siteblock.ini
@@ -1,0 +1,4 @@
+[SiteBlock]
+enable=true
+warningmsg=접속하는 사이트가 유해사이트로 등록되어있습니다.사용자정보, 접속URL, 접속시간이 보고됩니다.
+warningurl=http://warning.or.kr
