+++ hamonize-agent/etc/systemd/system/hamonize-agent.service
@@ -1,0 +1,27 @@
+[Unit]
+Description=hamonize-agent
+After=network.target
+
+[Service]
+Type=simple
+#Restart=always
+Restart=on-failure
+
+StartLimitInterval=60
+StartLimitBurst=5
+
+#Environment="DISPLAY=:0"
+
+WorkingDirectory=/usr/share/hamonize-agent
+ExecStart=/usr/share/hamonize-agent/hamonize-agent
+
+# make sure log directory exists and owned by syslog
+PermissionsStartOnly=true
+StandardOutput=syslog
+StandardError=syslog
+SyslogIdentifier=hamonize-agent
+
+[Install]
+WantedBy=multi-user.target
+
+
