--- hamonize-agent/exebuild_command.txt
@@ -1,7 +1,0 @@
-./node_modules/.bin/electron-packager ./ TestApp --platform-win43 \
---out dist \
---prune
-
-
-sudo ./node_modules/.bin/electron-packager ./ agent.AppImage --platform=linux --arch x64 --out dist --prune
-
