--- hamonize-agent/inject.js
@@ -1,7 +1,0 @@
-// node api 
-const  { ipcRenderer }  =  require ( 'electron' )
-
-ipcRenderer . on ( 'getContent' ,  function () { 
-  var  content  =  document . getElementsById ( "content" ); 
-  ipcRenderer . sendToHost ( 'getContent' ,  content . innerText ); 
-});
