--- hamonize-agent/package.json
@@ -1,75 +1,0 @@
-{
-  "name": "electron-quick-start",
-  "version": "1.0.0",
-  "description": "A minimal Electron application",
-  "main": "main.js",
-  "scripts": {
-    "start": "nodejs main.js",
-    "clean": "rm -rf dist",
-    "build": "pkg .  --targets node8-linux-x64"
-  },
-  "bin": {
-    "meshcentral": "./main.js",
-    "app": "."
-  },
-  "pkg": {
-    "assets": [
-      "public/**/*",
-      "bin/**/*",
-      "shell/*",
-      "*.sh"
-    ]
-  },
-  "repository": "https://github.com/electron/electron-quick-start",
-  "keywords": [
-    "Electron",
-    "quick",
-    "start",
-    "tutorial",
-    "demo"
-  ],
-  "author": "GitHub",
-  "license": "CC0-1.0",
-  "devDependencies": {
-    "electron": "^4.0.4",
-    "electron-builder": "^20.41.0",
-    "electron-packager": "^13.1.1"
-  },
-  "dependencies": {
-    "@actualwave/promised-timeout": "0.0.6",
-    "@andycasen/withtimeout": "^1.0.0",
-    "async-polling": "^0.2.1",
-    "create-file": "^1.0.1",
-    "date-utils": "^1.2.21",
-    "delay": "^4.4.1",
-    "dialog": "^0.3.1",
-    "electron-installer-squirrel-windows": "^1.3.0",
-    "electron-ipc-promise": "^0.1.7",
-    "electron-log": "^3.0.5",
-    "electron-widevinecdm": "^7.0.0",
-    "electron-window-state": "^5.0.3",
-    "fetch": "^1.1.0",
-    "fs": "0.0.1-security",
-    "ip": "^1.1.5",
-    "json-parser": "^1.1.5",
-    "node-cron": "^2.0.3",
-    "node-fetch": "^2.6.1",
-    "node-pre-gyp": "^0.17.0",
-    "node-schedule": "^1.3.3",
-    "ora": "^3.4.0",
-    "ping": "^0.2.3",
-    "pkg": "^4.4.9",
-    "poller": "0.0.5",
-    "precise": "^1.1.0",
-    "redux-electron-ipc": "^1.2.0",
-    "request": "^2.88.2",
-    "tenacious-fetch": "^2.3.0",
-    "time-counter": "^1.0.0",
-    "timer-stopwatch": "^0.2.0",
-    "winston": "^3.3.3",
-    "winston-daily-rotate-file": "^3.9.0",
-    "write": "^1.0.3",
-    "xml-js": "^1.6.11",
-    "xml2js": "^0.4.23"
-  }
-}
