+++ hamonize-agent/src/logger.js
