+++ hamonize-agent/src/shell/deviceResult.sh
