+++ hamonize-agent/src/usb-lockdown/30-usb-lockdown.rules
@@ -1,0 +1,64 @@
+# 장치를 사용할 수 없도록 하기 위해서는 /sys$devpath/authorized 값을 0으로 변경
+# $devpath : 현재 검색된 장치의 경로
+# 
+# 차단 메시지를 보이고 싶은 경우에는 아래 내용을 추가
+# ENV{DISPLAY}=":0.0", RUN+="/usr/local/bin/usb-unauthorized"
+# 
+# 디버깅 : tail -f /tmp/usb-lockdown.log
+
+# https://usb.org/defined-class-codes
+# https://www.beyondlogic.org/usbnutshell/usb5.shtml
+# 
+# Authorizing (or not) your USB devices to connect to the system
+# https://www.kernel.org/doc/Documentation/usb/authorization.txt
+# 
+# USB 장치의 유형으로 제어를 하는 경우
+# Ref : https://usb.org/defined-class-codes
+# Ref : https://www.beyondlogic.org/usbnutshell/usb5.shtml
+
+# 장치의 유형 (클래스가 Device인지 Interface인지 사용에 주의)
+# 00h : Use class information in the Interface Descriptors      Device
+# 01h : Audio                                                   Interface
+# 02h : Communications and CDC Control                          Both
+# 03h : HID (Human Interface Device)                            Interface       (키보드, 마우스)
+# 05h : Physical                                                Interface
+# 06h : Image                                                   Interface       (핸드폰 사진 전송 PTP)
+# 07h : Printer                                                 Interface       (프린터)
+# 08h : Mass Storage                                            Interface       (USB 저장장치)
+# 09h : Hub                                                     Device          (USB 멀티 허브)
+# 0Ah : CDC-Data                                                Interface
+# 0Bh : Smart Card                                              Interface
+# 0Dh : Content Security                                        Interface
+# 0Eh : Video                                                   Interface       (웹캠, 디지털 캠코더, 비디오 스트리밍 장치)
+# 0Fh : Personal Healthcare                                     Interface
+# 10h : Audio/Video Devices                                     Interface
+# 11h : Billboard Device Class                                  Device
+# 12h : USB Type-C Bridge Class                                 Interface
+# DCh : Diagnostic Device                                       Both
+# E0h : Wireless Controller                                     Interface
+# EFh : Miscellaneous                                           Both
+# FEh : Application Specific                                    Interface
+# FFh : Vendor Specific                                         Interface       (USB 무선 랜카드, USB 유선 랜카드)
+
+
+# 장치별로 전체 차단 (bInterfaceClass 기준으로 차단)
+# 해당 클래스는 USB가 꽂혔을 때 생긴다.
+ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="01", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+#ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="03", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="05", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="06", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+#ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="07", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="08", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="0a", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="0b", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="0d", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="0e", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="0f", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="10", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="12", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="e0", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="ef", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+ACTION=="add", SUBSYSTEMS=="usb", ENV{DEVTYPE}=="usb_interface", ATTRS{bInterfaceClass}=="ff", RUN="/usr/local/bin/usb-lockdown-work lockdown"
+
+
+### Center Add Vendor_ID PRODUCT_ID ###
