+++ hamonize-agent/src/usb-lockdown/usb-unauthorized
@@ -1,0 +1,35 @@
+#!/bin/bash
+
+LOGFILE="/tmp/usb-lockdown.log"
+echo "=== Usable Global Variable Info ===" | tee -a $LOGFILE
+set 2>&1 | tee -a $LOGFILE
+
+echo "=== Patameters from *.rules file ===" | tee -a $LOGFILE
+echo $1  | tee -a $LOGFILE
+
+echo "=== Device PATH INFO ===" | tee -a $LOGFILE
+echo "Lock down PATH : "${LOCKDOWNPATH} | tee -a $LOGFILE
+
+xcheck=`tty | cut -d '/' -f3`
+if [ $xcheck != "pts" ]; then
+    echo "Current Mode is CLI : This Program must be run under GUI" 2>&1 | tee -a $LOGFILE
+    exit
+fi
+
+# Real Username and UID
+RUID=$(who | awk 'FNR == 1 {print $1}')
+RUSER_UID=$(id -u ${RUID})
+
+authorized=$(cat /sys${DEVPATH}/authorized)
+authorized_default=$(cat /sys${DEVPATH}/authorized_default)
+
+echo "/sys${DEVPATH}/authorized= $authorized"  | tee -a $LOGFILE
+echo "/sys${DEVPATH}/authorized_default= $authorized_default"  | tee -a $LOGFILE
+
+if [[ $LANG == "ko_KR.UTF-8" || -z $LANG ]]; then
+    sudo -u ${RUID} DBUS_SESSION_BUS_ADDRESS="unix:path=/run/user/${RUSER_UID}/bus" zenity --error --width=400 --height=100 --text "허용된 USB 만 사용 가능합니다.\n\n관리자에 의해서 이 USB 장치 사용이 금지되었습니다.\n\n관리자에게 >문의하세요.\n\n - Name : ${DEVNAME}\n - Path : ${DEVPATH}\n - Type : ${DEVTYPE}\n - Driver : ${DRIVER}\n - Model : ${ID_VENDOR} ${ID_MODEL}\n - Vendor:Model ${ID_VENDOR_ID}:${ID_MODEL_ID}\n\n - authorized location : /sys${DEVPATH}/authorized\n - authorized : ${authorized}"
+else
+    sudo -u ${RUID} DBUS_SESSION_BUS_ADDRESS="unix:path=/run/user/${RUSER_UID}/bus" zenity --error --width=400 --height=100 --text "Only allowed USBs can be used.\n\nThis device is prohibited by the administrator.\n\nPlease contact the administrator.\n\n - Name : ${DEVNAME}\n - Path : ${DEVPATH}\n - Type : ${DEVTYPE}\n - Driver : ${DRIVER}\n - Model : ${ID_VENDOR} ${ID_MODEL}\n - Vendor:Model ${ID_VENDOR_ID}:${ID_MODEL_ID}\n\n - authorized location : /sys${DEVPATH}/authorized\n - authorized : ${authorized}"
+fi
+
+exit
